requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
jinja2==3.1.3
python-multipart==0.0.9
aiofiles==23.2.1
pyyaml==6.0.1
